fixes emscripten build bug by ensuring conan_paths.cmake is found
There was an issue when CIL ran `package all`, which causes a distinct
directory to be used for "install", so the conan_paths.cmake file isn't
in the same place as all the binaries. My clunky gross solution is to
have Conan tell the Cmake file the location of this explicitly (maybe
passing the toolchain would work better, but that seemed like too big of
a change- I still have very faint hope I can make this work with Cget
as well).